+++ src/extensions/rules/rulesext.cpp
@@ -113,7 +113,8 @@ RulesClassExtension::RulesClassExtension(const RulesClass* this_ptr) :
     BaseUnit(),
     Diff(),
     PlayerNormal(),
-    IsHasPlayerNormal(false)
+    IsHasPlayerNormal(false),
+    PersistTagsOnAIDeploy(false)
 {
     /**
      *  Due to the changes made when addressing issues #632, 633, and 635, we
@@ -280,6 +281,7 @@ void RulesClassExtension::Object_CRC(CRCEngine &crc) const
     crc(IsCellTagsIgnoreStealth);
     crc(SelfHealingStep);
     crc(BaseUnit.Count());
+    crc(PersistTagsOnAIDeploy);
 }
 
 
@@ -715,6 +717,7 @@ bool RulesClassExtension::General(CCINIClass &ini)
     AbandonTargetEscortRange = ini.Get_Lepton(GENERAL, "AbandonTargetEscortRange", AbandonTargetEscortRange);
     IsFreeRadarOnLowPower = ini.Get_Bool(GENERAL, "FreeRadarOnLowPower", IsFreeRadarOnLowPower);    
     IsCellTagsIgnoreStealth = ini.Get_Bool(GENERAL, "CellTagsIgnoreStealth", IsCellTagsIgnoreStealth);
+    PersistTagsOnAIDeploy = ini.Get_Bool(GENERAL, "PersistTagsOnAIDeploy", PersistTagsOnAIDeploy);
 
     /**
      *  Allow replacing any signle movement zone with a copy of RA2's water MZone.